+++ test/files/files.spec.ts
@@ -4,9 +4,10 @@ import { Cassettes } from 'mocha-cassettes';
 import { LokaliseApi } from '../../src/lokalise/lokalise';
 
 describe('Files', function () {
-  const cassette = new Cassettes('./test/cassettes');
-  const lokaliseApi = new LokaliseApi({apiKey: process.env.API_KEY});
-  const project_id = '803826145ba90b42d5d860.46800099';
+  const cassette = new Cassettes('./test/cassettes')
+  const lokaliseApi = new LokaliseApi({apiKey: process.env.API_KEY})
+  const project_id = '803826145ba90b42d5d860.46800099'
+  const process_id = '3b943469e6b3e324b5bdad639b122a623e6e7a1a'
 
   cassette.createTest('list', async () => {
     const files = await lokaliseApi.files.list({ project_id: project_id});
@@ -26,6 +27,35 @@ describe('Files', function () {
     expect(response.file).to.eq('test1.json');
     expect(response.result['inserted']).to.eq(3);
   }).register(this);
+
+  cassette.createTest('upload asynchronous', async () => {
+    const data = 'ewogICAgImZydWl0IjogIkFwcGxlIiwKICAgICJzaXplIjogIkxhcmdlIiwKICAgICJjb2xvciI6ICJSZWQiCn0=';
+    const process = await lokaliseApi.files.upload(project_id,
+      {data: data, filename: 'test_async.json', lang_iso: 'en', queue: true}
+    )
+
+    expect(process.process_id).to.eq(process_id)
+    expect(process.type).to.eq('file-import')
+    expect(process.status).to.eq('queued')
+    expect(process.message).to.eq('')
+    expect(process.created_by).to.eq('20181')
+    expect(process.created_by_email).to.eq('bodrovis@protonmail.com')
+    expect(process.created_at).to.eq('2020-05-13 11:24:37 (Etc/UTC)')
+    expect(process.created_at_timestamp).to.eq(1589369077)
+    expect(process.url).to.eq('/api2/projects/803826145ba90b42d5d860.46800099/processes/file-import/3b943469e6b3e324b5bdad639b122a623e6e7a1a')
+  }).register(this);
+
+  cassette.createTest('upload asynchronous re-check', async () => {
+    const process = await lokaliseApi.queuedProcesses.
+      getDetailed(process_id, { project_id: project_id })
+
+    expect(process.process_id).to.eq(process_id)
+    expect(process.status).to.eq('finished')
+    expect(process.files.length).to.eq(1)
+    const file = process.files[0]
+    expect(file.name_original).to.eq('test_async.json')
+    expect(file.word_count_total).to.eq(3)
+  }).register(this)
 
   cassette.createTest('download', async () => {
     const response = await lokaliseApi.files.download(project_id,